+++ src/CoordinateValueObject.php
@@ -4,26 +4,18 @@ declare(strict_types=1);
 
 namespace Masfernandez\ValueObject;
 
-use LongitudeOne\Spatial\Exception\InvalidValueException;
 use LongitudeOne\Spatial\PHP\Types\Geometry\Point;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints;
 
 abstract class CoordinateValueObject extends ValueObject
 {
-    private Point $point;
+    private readonly Point $point;
 
-    /**
-     * @throws ValueObjectException
-     */
     public function __construct(float $x, float $y)
     {
-        parent::__construct([$x, $y]);
-        try {
-            $this->point = new Point($x, $y);
-        } catch (InvalidValueException $e) {
-            throw new ValueObjectException($e->getMessage(), $e->getCode(), $e);
-        }
+        parent::__construct(['x' => $x, 'y' => $y]);
+        $this->point = new Point($x, $y);
     }
 
     /**
@@ -43,6 +35,9 @@ abstract class CoordinateValueObject extends ValueObject
         ];
     }
 
+    /**
+     * @return float[]
+     */
     public function value(): array
     {
         return $this->point->toArray();
@@ -50,6 +45,6 @@ abstract class CoordinateValueObject extends ValueObject
 
     public function __toString(): string
     {
-        return $this->point->toJson();
+        return "[{$this->point->getX()},{$this->point->getY()}]";
     }
 }