+++ src/Exception/EnumValueObjectException.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Masfernandez\ValueObject;
+namespace Masfernandez\ValueObject\Exception;
 
 class EnumValueObjectException extends ValueObjectException
 {
-
 }